+++ example/lib/examples/extended_example/helper_widgets/helper_widgets.dart
@@ -1,0 +1,1 @@
+export 'scroll_ijector.dart';