+++ example/lib/examples/extended_example/helper_widgets/scroll_ijector.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+import 'package:group_button/group_button.dart';
+
+class ScrollIjector extends StatelessWidget {
+  const ScrollIjector({
+    Key key,
+    @required this.child,
+    @required this.groupingType,
+  }) : super(key: key);
+
+  final Widget child;
+  final GroupingType groupingType;
+
+  @override
+  Widget build(BuildContext context) {
+    if (groupingType == GroupingType.row) {
+      return SingleChildScrollView(
+        scrollDirection: Axis.horizontal,
+        child: child,
+      );
+    }
+    return child;
+  }
+}