+++ lib/src/group_custom_button.dart
@@ -6,12 +6,12 @@ class GroupCustomButton extends StatelessWidget {
     required this.onPressed,
     required this.text,
     required this.isSelected,
+    this.selectedBorderColor,
+    this.unselectedBorderColor,
     this.selectedTextStyle,
     this.unselectedTextStyle,
     this.selectedColor,
     this.unselectedColor,
-    required this.selectedBorderColor,
-    required this.unselectedBorderColor,
     this.borderRadius,
     this.selectedShadow,
     this.unselectedShadow,
@@ -26,8 +26,8 @@ class GroupCustomButton extends StatelessWidget {
   final TextStyle? unselectedTextStyle;
   final Color? selectedColor;
   final Color? unselectedColor;
-  final Color selectedBorderColor;
-  final Color unselectedBorderColor;
+  final Color? selectedBorderColor;
+  final Color? unselectedBorderColor;
   final BorderRadius? borderRadius;
   final List<BoxShadow>? selectedShadow;
   final List<BoxShadow>? unselectedShadow;
@@ -36,6 +36,7 @@ class GroupCustomButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final theme = Theme.of(context);
     return Container(
       height: height,
       width: width,
@@ -47,17 +48,8 @@ class GroupCustomButton extends StatelessWidget {
         onPressed: onPressed,
         style: ButtonStyle(
           elevation: MaterialStateProperty.all<double>(0.0),
-          backgroundColor: isSelected
-              ? MaterialStateProperty.all<Color?>(selectedColor)
-              : MaterialStateProperty.all<Color?>(unselectedColor),
-          shape: MaterialStateProperty.all<OutlinedBorder>(
-            RoundedRectangleBorder(
-              borderRadius: borderRadius ?? BorderRadius.circular(30),
-              side: BorderSide(
-                color: isSelected ? selectedBorderColor : unselectedBorderColor,
-              ),
-            ),
-          ),
+          backgroundColor: _getBackGroundColor(theme),
+          shape: _buildShape(),
         ),
         child: Text(
           text,
@@ -65,5 +57,51 @@ class GroupCustomButton extends StatelessWidget {
         ),
       ),
     );
+  }
+
+  MaterialStateProperty<Color?> _getBackGroundColor(ThemeData theme) {
+    final themePrimaryColor = theme.buttonTheme.colorScheme?.primary;
+    //TODO: implement
+    // final themeSecondaryColor = theme.buttonTheme.colorScheme?.secondary;
+
+    Color? selectedColorBuffer = selectedColor;
+    Color? unselectedColorBuffer = unselectedColor;
+
+    selectedColorBuffer ??= themePrimaryColor;
+    selectedColorBuffer ??= Colors.black;
+
+    unselectedColorBuffer ??= Colors.white;
+
+    final color = isSelected ? selectedColorBuffer : unselectedColorBuffer;
+    return MaterialStateProperty.all<Color?>(color);
+  }
+
+  MaterialStateProperty<OutlinedBorder>? _buildShape() {
+    final color = isSelected
+        ? selectedBorderColor ?? unselectedBorderColor
+        : unselectedBorderColor;
+    if (borderRadius != null) {
+      return MaterialStateProperty.all<OutlinedBorder>(
+        RoundedRectangleBorder(
+          borderRadius: borderRadius!,
+          side: buildBorderSide(color),
+        ),
+      );
+    } else {
+      return MaterialStateProperty.all<OutlinedBorder>(
+        RoundedRectangleBorder(
+          side: buildBorderSide(color),
+        ),
+      );
+    }
+  }
+
+  BorderSide buildBorderSide(Color? color) {
+    if (color != null) {
+      return BorderSide(
+        color: color,
+      );
+    }
+    return BorderSide.none;
   }
 }