+++ lib/src/utils/grouping_type.dart
@@ -1,0 +1,5 @@
+enum GroupingType {
+  wrap,
+  column,
+  row,
+}